tools/hydra-test-tool/pyproject.toml
[project]
name = "hydra-test-tool"
version = "0.1.0"
description = "Minimal Hydra tool for testing ExperimentStash integration"
requires-python = ">=3.8"
dependencies = [
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]